+++ src/Relation/BelongsTo.php
@@ -40,7 +40,7 @@ final class BelongsTo extends AbstractRelation
             return null;
         }
 
-        return $query->first();
+        return $query->oneOrNull();
     }
 
     public function save(Model $model): Model