+++ src/Relation/Polymorphic.php
@@ -88,7 +88,7 @@ final class Polymorphic extends AbstractRelation
             return null;
         }
 
-        return $query->first();
+        return $query->oneOrNull();
     }
 
     public function save(Model $model): Model