+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - The new `ListenerException` class can be thrown from event listeners as a shortcut to stop events.
 - A `deleted` property has been added to soft deleted models.
 - The `Query::one()` method finds exactly one model or throws an exception.
+- The `Query::oneOrNull()` method finds exactly one model or returns null.
 - Capture the options of the failing validation rule for more accurate error messages.
 
 ### Changed